requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
yfinance>=0.2.36
gTTS>=2.4.0
flask>=3.0.0
schedule>=1.2.0
requests>=2.31.0
